Various changes to get it running again with latest libopenmpt and ChromaPrint, eliminate Qt deprecation warnings

+++ Mod Library/audioplayer.h
@@ -58,7 +58,7 @@ public slots:
 				break;
 			}
 			const float * const buffers [2] = { left.data(), right.data() };
-			Pa_WriteStream(stream, buffers, count);
+			Pa_WriteStream(stream, buffers, static_cast<unsigned long>(count));
 		}
 		Pa_StopStream(stream);
 		Pa_CloseStream(stream);
